+++ CHANGELOG.md
@@ -8,10 +8,16 @@ and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2
 ## v2.12.0-pre.3
 ### Added
 - Added Pip Overrides section to the package install view, allowing you to override the default pip packages for a package
+### Changed
+- Show better error messsage for early access model downloads
+- Updated torch version for a1111 on mac
+- Checkpoints tab now shows "image hidden" for images that are hidden by the NSFW filter
 ### Fixed
 - Fixed issue where some Inference-generated images would be saved with the bottom missing
 - Fixed CivitAI Browser page scroll refresh not ordering models correctly
 - Fixed missing fbgemm.dll errors when using latest torch with certain packages
+- Fixed issue where ApproxVAE models would show up in the VAE folder
+- Checkpoints tab will no longer try to browse directories it can't access
 
 ## v2.12.0-pre.2
 ### Added